# out/src/Johnfg2610.Stockit/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "johnfg2610@gmail.com",
      "name" : "john",
      "url" : "https://www.github.com/johnfg2610/stockit"
    },
    "description" : "Used to access inventory counts as well as brands and catalog items",
    "license" : {
      "name" : "ISC",
      "url" : "https://github.com/johnfg2610/stockit/blob/master/LICENSE"
    },
    "title" : "Inventory",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "security" : [ {
    "jwt-auth" : [ ]
  } ],
...
  "paths" : {
    "/inventory/{catalogItemId}" : {
      "get" : {
        "description" : "Gets the amount of items avalible in inventory of the requested catalogItem",
        "operationId" : "get-inventory-catalogItemId",
        "parameters" : [ {
          "description" : "The catalog item to modify",
          "explode" : false,
          "in" : "path",
          "name" : "catalogItemId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example-1" : {
                    "value" : 32
                  }
                },
                "schema" : {
                  "default" : 0,
                  "description" : "The amount of the selected product avalible",
                  "type" : "integer"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Get the current amount of the item available",
        "tags" : [ "inventory" ]
      },
      "patch" : {
        "description" : "Updates the amount of a item",
        "operationId" : "patch-inventory-catalogItemId",
        "parameters" : [ {
          "description" : "The catalog item to modify",
          "explode" : false,
          "in" : "path",
          "name" : "catalogItemId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/inline_object",
          "content" : {
            "application/json" : {
              "examples" : {
                "example-1" : {
                  "value" : {
                    "amountIncrease" : 35
                  }
                }
              },
              "schema" : {
                "properties" : {
                  "amountChange" : {
                    "type" : "integer"
                  }
                },
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Update the amount of a item",
        "tags" : [ "inventory" ]
      }
    },
...
  },
  "components" : {
    "requestBodies" : {
      "inline_object" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/inline_object"
            }
          }
        }
      }
    },
    "schemas" : {
...
      "Error" : {
        "description" : "Used to describe a error that happened during the execution of a service",
        "properties" : {
          "code" : {
            "type" : "number"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "title" : "Error",
        "type" : "object"
      },
      "inline_object" : {
        "properties" : {
          "amountChange" : {
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "inline_response_201" : {
        "example" : {
          "newAmount" : 0
        },
        "properties" : {
          "newAmount" : {
            "type" : "integer"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "jwt-auth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}